PIM-6353: Renames the product indexer service
