+++ src/Pim/Bundle/ApiBundle/tests/integration/Controller/Product/DeleteProductIntegration.php
@@ -22,7 +22,7 @@ class DeleteProductIntegration extends AbstractProductTestCase
         $this->assertCount(4, $this->get('pim_catalog.repository.product')->findAll());
 
         $fooProduct = $this->get('pim_catalog.repository.product')->findOneByIdentifier('foo');
-        $this->get('pim_catalog.elasticsearch.product_indexer')->index($fooProduct);
+        $this->get('pim_catalog.elasticsearch.indexer.product')->index($fooProduct);
         $client->request('DELETE', 'api/rest/v1/products/foo');
 
         $response = $client->getResponse();
