+++ src/Pim/Bundle/CatalogBundle/Resources/config/removers.yml
@@ -82,7 +82,7 @@ services:
         arguments:
             - '@pim_catalog.query.product_query_builder_factory'
             - '@pim_catalog.object_manager.product'
-            - '@pim_catalog.elasticsearch.product_indexer'
+            - '@pim_catalog.elasticsearch.indexer.product'
             - 'pim_catalog_completeness'
 
     pim_catalog.remover.base_options_resolver:
