+++ src/Pim/Bundle/DataGridBundle/Resources/config/mass_actions.yml
@@ -72,7 +72,7 @@ services:
     pim_datagrid.extension.mass_action.handler.product_delete:
         class: '%pim_datagrid.extension.mass_action.handler.product_delete.class%'
         arguments:
-            - '@pim_catalog.elasticsearch.product_indexer'
+            - '@pim_catalog.elasticsearch.indexer.product'
         parent: pim_datagrid.extension.mass_action.handler.delete
         tags:
             - { name: pim_datagrid.extension.mass_action.handler, alias: product_mass_delete }
